backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
opencv-python-headless
pillow
pandas
numpy
httpx
python-multipart
scikit-learn
python-dotenv
supabase
python-jose[cryptography]
passlib[bcrypt]
